fix: devtools evaluate crashes on single thread mode
