+++ test/inspector/test-inspect.js
@@ -25,34 +25,6 @@ promise_test(async () => {
 
   await client.callMethod('Debugger.enable');
 
-  let pausedFuture = once(client, 'Debugger.paused');
-  await client.callMethod('Runtime.runIfWaitingForDebugger');
-  await pausedFuture;
-
-  pausedFuture = once(client, 'Debugger.paused');
-  await client.callMethod('Debugger.resume');
-  await pausedFuture;
-
-  const { result } = await client.callMethod('Runtime.evaluate', {
-    expression: 'foo',
-  });
-  assert_equals(result.type, 'string');
-  assert_equals(result.value, 'bar');
-
-  // close client;
-  client.reset();
-  const [ exitCode ] = await once(cp, 'exit');
-  assert_equals(exitCode, 0);
-}, 'inspect commands');
-
-promise_test(async () => {
-  const cp = await run(inspectScript);
-
-  const client = new Client();
-  await client.connect(9229, 'localhost');
-
-  await client.callMethod('Debugger.enable');
-
   const pausedFuture = once(client, 'Debugger.paused');
   await client.callMethod('Runtime.runIfWaitingForDebugger');
   const [ params ] = await pausedFuture;
